+++ packages/waypoint/src/provider.ts
@@ -25,7 +25,7 @@ import { openPopup } from "./utils/popup"
 import { getStorage, removeStorage, setStorage, STORAGE_ADDRESS_KEY } from "./utils/storage"
 import { validateIdAddress } from "./utils/validate-address"
 
-export type RoninWaypointWalletOpts = AuthorizeOpts & {
+export type WaypointProviderOpts = AuthorizeOpts & {
   chainId: number
 }
 
@@ -34,19 +34,21 @@ export type RoninWaypointWalletOpts = AuthorizeOpts & {
  *
  * This provider is designed to easily integrate with Ronin Waypoint.
  *
- * Use `create` function to create a new instance.
+ * Use `create` function to create a new instance.ddress"
+import { EventEmitter } from "events"
+import { A, ChainDisconnectedError, Client, createClient, EIP1193Parameters, http, ProviderDisconnectedError, toHex, UnauthorizedProviderError } from "viem
  *
  * @example
- * import { RoninWaypointWallet } from "@sky-mavis/waypoint"
+ * import { WaypointProvider } from "@sky-mavis/waypoint"
  *
- * const idWalletProvider = RoninWaypointWallet.create({
+ * const idWalletProvider = WaypointProvider.create({
  *  clientId: "YOUR_CLIENT_ID",
  *  chainId: ronin.chainId,
  * })
  */
-export class RoninWaypointWallet extends EventEmitter implements Eip1193Provider {
+export class WaypointProvider extends EventEmitter implements Eip1193Provider {
   private readonly clientId: string
-  private readonly idOrigin: string
+  private readonly waypointOrigin: string
   private readonly redirectUrl: string
   private readonly scopes: Scope[]
 
@@ -56,23 +58,23 @@ export class RoninWaypointWallet extends EventEmitter implements Eip1193Provider
   private readonly viemClient: Client
   private readonly communicateHelper: CommunicateHelper
 
-  protected constructor(options: RoninWaypointWalletOpts) {
+  protected constructor(options: WaypointProviderOpts) {
     super()
 
     const {
       clientId,
       chainId,
       scopes = [],
-      idOrigin = RONIN_WAYPOINT_ORIGIN_PROD,
+      waypointOrigin = RONIN_WAYPOINT_ORIGIN_PROD,
       redirectUrl = typeof window !== "undefined" ? window.location.origin : "",
     } = options
 
     this.clientId = clientId
-    this.idOrigin = idOrigin
+    this.waypointOrigin = waypointOrigin
     this.redirectUrl = redirectUrl
     this.chainId = chainId
     this.scopes = this.addDefaultScopes(scopes)
-    this.communicateHelper = new CommunicateHelper(idOrigin)
+    this.communicateHelper = new CommunicateHelper(waypointOrigin)
     this.viemClient = this.createViemClient(chainId)
   }
 
@@ -105,22 +107,22 @@ export class RoninWaypointWallet extends EventEmitter implements Eip1193Provider
   }
 
   /**
-   * Creates a new RoninWaypointWallet instance.
+   * Creates a new WaypointProvider instance.
    *
-   * @param options Options for RoninWaypointWallet.
+   * @param options Options for WaypointProvider.
    *
-   * @returns RoninWaypointWallet instance.
+   * @returns WaypointProvider instance.
    *
    * @example
-   * import { RoninWaypointWallet } from "@sky-mavis/waypoint"
+   * import { WaypointProvider } from "@sky-mavis/waypoint"
    *
-   * const idWalletProvider = RoninWaypointWallet.create({
+   * const idWalletProvider = WaypointProvider.create({
    *  clientId: "YOUR_CLIENT_ID",
    *  chainId: ronin.chainId,
    * })
    */
-  public static create = (options: RoninWaypointWalletOpts) => {
-    return new RoninWaypointWallet(options)
+  public static create = (options: WaypointProviderOpts) => {
+    return new WaypointProvider(options)
   }
 
   private getIdAddress = () => {
@@ -144,10 +146,10 @@ export class RoninWaypointWallet extends EventEmitter implements Eip1193Provider
    * @returns The access token and address.
    */
   connect = async () => {
-    const { idOrigin, clientId, redirectUrl, scopes, communicateHelper, chainId } = this
+    const { waypointOrigin, clientId, redirectUrl, scopes, communicateHelper, chainId } = this
 
     const authData = await communicateHelper.sendRequest<IdResponse>(state =>
-      openPopup(`${idOrigin}/client/${clientId}/authorize`, {
+      openPopup(`${waypointOrigin}/client/${clientId}/authorize`, {
         state,
         redirect: redirectUrl,
         origin: window.location.origin,
@@ -206,7 +208,7 @@ export class RoninWaypointWallet extends EventEmitter implements Eip1193Provider
   request = async <ReturnType = unknown>(args: EIP1193Parameters<RoninWaypointRequestSchema>) => {
     const {
       clientId,
-      idOrigin,
+      waypointOrigin,
       communicateHelper,
       chainId,
       viemClient,
@@ -241,7 +243,7 @@ export class RoninWaypointWallet extends EventEmitter implements Eip1193Provider
           params,
           expectAddress,
           clientId,
-          idOrigin,
+          waypointOrigin,
           communicateHelper,
         }) as ReturnType
       }
@@ -254,7 +256,7 @@ export class RoninWaypointWallet extends EventEmitter implements Eip1193Provider
           chainId,
           expectAddress,
           clientId,
-          idOrigin,
+          waypointOrigin,
           communicateHelper,
         }) as ReturnType
       }
@@ -267,7 +269,7 @@ export class RoninWaypointWallet extends EventEmitter implements Eip1193Provider
           chainId,
           expectAddress,
           clientId,
-          idOrigin,
+          waypointOrigin,
           communicateHelper,
         }) as ReturnType
       }